Move chain of trust tests into chain_of_test.go
